Add event sheet export check
Event sheets can no longer be exported as long as the participation is not finished.
This changle only affects the UI, because technically it is not wrong
to export the event sheets even before the participation is finished.
But the sheets would be empty which makes no sense from a user stand point.